requirements.txt
pandas==2.1.1
numpy==1.26.0
pint==0.22
pyyaml==6.0.0
